fixed 5 piece auto
Such that all the intakes and outtakes work for the normal one not the under the stage one, but the final piece does not outtake because the indexer does not run for some reason. Also, important to note that this was all done with the arm code commented out, so this will need to be fixed. And a register command was added where the outtake can run the whole time, but we did not have time to test that one. Running 5 piece in ~20 seconds with 3.25 velocity and acceleration, fixed the robot dimensions with the bumper. Also the 4 piece works in the 15 seconds time limit. Tried to make a stop/wait command, doesn't work. Also do not make folders in path planner.