working 5.5 piece 5 piece
and path for 6 piece, also test choreo source side 3 piece which was not that bad. Did cycle very well and amp working. Podium shoots while moving really slow completely working on the move and podium stationary working almost 100% of the time.  Middle of the stage not the accurate distance and does not work at all. Preparing shooter to shoot seems to be a little bit faster as well. Lot of stuff left to do still.